refactor: replace `gzgov` by `gz` and change doc accordingly
