+++ lib/v2/gov/router.js
@@ -38,6 +38,7 @@ module.exports = function (router) {
     router.get('/beijing/kw/:channel', require('./beijing/kw/index'));
     router.get('/guangdong/tqyb/tfxtq', require('./guangdong/tqyb/tfxtq'));
     router.get('/guangdong/tqyb/sncsyjxh', require('./guangdong/tqyb/sncsyjxh'));
+    router.get('/gz/:channel/:category', require('./gz/index'));
     router.get('/hebei/czt/xwdt/:category?', require('./hebei/czt'));
     router.get('/huizhou/zwgk/:category?', require('./huizhou/zwgk/index'));
     router.get('/hunan/changsha/major-email', require('./hunan/changsha/major-email'));
@@ -51,5 +52,4 @@ module.exports = function (router) {
     router.get('/sichuan/deyang/govpulicinfo/:countyName', require('./sichuan/deyang/govpulicinfo'));
     router.get('/taiyuan/rsj/:caty/:page?', require('./taiyuan/rsj'));
     router.get('/xuzhou/hrss/:category?', require('./xuzhou/hrss'));
-    router.get('/gzgov/:channel/:category', require('./gzgov/index'));
 };
